+++ include/openrave/viewer.h
@@ -82,16 +82,17 @@ public:
     /// \brief reset the camera depending on its mode
     virtual void UpdateCameraTransform() { throw openrave_exception("ViewerBase::UpdateCameraTransform not implemented",ORE_NotImplemented); }
 
-    /// Renders a 24bit RGB image of dimensions width and height from the current scene. The camera
-    /// is meant to show the underlying OpenRAVE world as a robot would see it, so all graphs
-    /// rendered with the plotX and drawX functions are hidden.
-    /// \param memory the memory where the image will be stored at, has to store 3*width*height
-    /// \param width width of the image
-    /// \param height height of the image
-    /// \param t the rotation and translation of the camera. Note that z is treated as the front of the camera!
-    ///        So all points in front of the camera have a positive dot product with its direction.
-    /// \param KK 4 values such that the intrinsic matrix can be reconstructed [pKK[0] 0 pKK[2]; 0 pKK[1] pKK[3]; 0 0 1];
-    virtual bool GetCameraImage(std::vector<uint8_t>& memory, int width, int height, const RaveTransform<float>& t, const SensorBase::CameraIntrinsics& KK) { throw openrave_exception("ViewerBase::GetCameraImage not implemented",ORE_NotImplemented); }
+    /** \brief Renders a 24bit RGB image of dimensions width and height from the current scene.
+
+        The camera is meant to show the underlying OpenRAVE world as a robot would see it, so all graphs
+        rendered with the plotX and drawX functions are hidden by default. Some viewers support the SetFiguresInCamera command to allow graphs to be also displayed.
+        \param memory the memory where the image will be stored at, has to store 3*width*height
+        \param width width of the image
+        \param height height of the image
+        \param t the rotation and translation of the camera. Note that +z is treated as the camera direction axis! So all points in front of the camera have a positive dot product with its +z direction.
+        \param intrinsics the intrinsic parameters of the camera defining FOV, distortion, principal point, and focal length. The focal length is used to define the near plane for culling.
+    */
+    virtual bool GetCameraImage(std::vector<uint8_t>& memory, int width, int height, const RaveTransform<float>& t, const SensorBase::CameraIntrinsics& intrinsics) { throw openrave_exception("ViewerBase::GetCameraImage not implemented",ORE_NotImplemented); }
 
     //@}